feat: agregar código único a solicitudes, mejorar gestión de estados y detalles de cierre

+++ reportes_mtto/solicitudes/views.py
@@ -4,6 +4,7 @@ from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from .forms import SolicitudForm
 from django.shortcuts import redirect
 from django.utils import timezone
+from django.contrib import messages
 # Create your views here.
 
 #Se crea vista para listar solicitud de mantenimiento activas
@@ -48,38 +49,48 @@ def detalle_solicitud(request, id):
     solicitud = get_object_or_404(Solicitud,id=id)
     return render(request, 'solicitud_detail.html', {'solicitud': solicitud})
 
+
 #Se crea vista para editar solicitud de mantenimiento, detectando cambios de estado y agregando comentarios de seguimiento
 def editar_solicitud(request, id):
     solicitud = get_object_or_404(Solicitud, id=id)
-    estado_anterior = solicitud.get_estado_display().capitalize() # Obtener la representación legible del estado antes de la actualización
     form = SolicitudForm(request.POST or None, instance=solicitud)
-    if request.method == 'POST' and form.is_valid():
-        solicitud_actualizada = form.save()
+    if request.method == 'POST':
         comentario = request.POST.get('comentario')
-        nuevo_estado = solicitud_actualizada.get_estado_display().capitalize()  # Obtener la representación legible del estado
-        texto_seguimiento = ''
-        # Detectar cambio de estado
-        if estado_anterior != nuevo_estado:
-            texto_seguimiento += (
+        if solicitud.estado == 'PENDIENTE':
+            if form.is_valid():
+                form.save()
+                return redirect('solicitudes:lista')
+        elif solicitud.estado == 'EN_PROCESO':
+            comentario = request.POST.get('comentario')
+            if comentario:
+                Seguimiento.objects.create(
+                    solicitud=solicitud,
+                    comentario=f'<strong>Comentario:</strong> {comentario}'
+                )
+            return redirect('solicitudes:lista')
+            
+                
+    return render(request, 'solicitud_form.html', {'form': form, 'solicitud': solicitud})
+
+#Se crea vista para boton de cambio de estado de pendiente a en proceso, detectando cambios de estado y agregando comentarios de seguimiento
+def cambiar_estado_en_proceso(request, id):
+    solicitud = get_object_or_404(Solicitud, id=id)
+    if solicitud.estado == 'PENDIENTE':
+        estado_anterior = solicitud.get_estado_display().capitalize()
+        solicitud.estado = 'EN_PROCESO'
+        solicitud.save()
+        nuevo_estado = solicitud.get_estado_display().capitalize()
+        Seguimiento.objects.create(
+            solicitud=solicitud,
+            comentario=(
                 f'<strong>Estado:</strong> '
                 f'{estado_anterior} → {nuevo_estado}<br>'
-            )
-            
-        #Comentario manual
-        if comentario:
-            texto_seguimiento += (
                 f'<strong>Comentario:</strong> '
-                f'{comentario} '
+                f'Cambio automático al iniciar proceso de mantenimiento.'
             )
-            
-        #Crea un solo comentario de seguimiento si hay cambios de estado o comentario manual
-        if texto_seguimiento:
-            Seguimiento.objects.create(
-                solicitud=solicitud,
-                comentario=texto_seguimiento
-            )
-        return redirect('solicitudes:lista')
-    return render(request, 'solicitud_form.html', {'form': form, 'solicitud': solicitud})
+        )
+    return redirect('solicitudes:solicitud', id=solicitud.id)
+
 
 #Se crea vista para cerrar solicitud de mantenimiento, detectando cambios de estado y agregando comentarios de seguimiento
 def cerrar_solicitud(request, id):
@@ -89,8 +100,15 @@ def cerrar_solicitud(request, id):
     if solicitud.estado == 'CERRADA':
         return redirect('solicitudes:detalle', id=solicitud.id)
     if request.method == 'POST':
-        trabajo_realizado = request.POST.get('trabajo_realizado')
-        observaciones_cierre = request.POST.get('observaciones_cierre')
+        trabajo_realizado = request.POST.get('trabajo_realizado', '').strip()
+        observaciones_cierre = request.POST.get('observaciones_cierre', '').strip()
+        
+        if not trabajo_realizado:
+            return render(request, 'solicitud_cerrar.html', {
+                'solicitud': solicitud,
+                'error': 'El trabajo realizado es obligatorio.'
+            })
+            
         # Actualiza datos de cierre
         solicitud.trabajo_realizado = trabajo_realizado
         solicitud.observaciones_cierre = observaciones_cierre
@@ -109,10 +127,24 @@ def cerrar_solicitud(request, id):
         return redirect('solicitudes:lista')
     return render(request,'solicitud_cerrar.html',{'solicitud': solicitud})
 
+
+#Se crea vista para detalle de cierre de solicitud de mantenimiento
+def detalle_cierre(request, id):
+    solicitud = get_object_or_404(Solicitud, id=id)
+    if solicitud.estado != 'CERRADA':
+        return redirect('solicitudes:detalle', id=solicitud.id)
+    return render(request, 'solicitud_cierre_detail.html', {'solicitud': solicitud})
+
 #Se crea vista para eliminar solicitud de mantenimiento
 def eliminar_solicitud(request, id):
     solicitud = get_object_or_404(Solicitud, id=id)
     if request.method == 'POST':
+        if solicitud.estado != 'PENDIENTE':
+            messages.error(
+                request,
+                'No se puede eliminar una orden en proceso o cerrada.'
+            )
+            return redirect('solicitudes:lista')
         solicitud.delete()
     return redirect('solicitudes:lista')
